# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Payment Gateway API
  version: "1.1.0"
  description: >
    USDT payment gateway supporting BEP20 (BNB Smart Chain) and TRC20 (Tron).
    Merchant endpoints authenticate with API key + HMAC signature; dashboard
    endpoints authenticate with JWT. The target chain is chosen per-payment via
    the optional `network` field (defaults to BEP20 for backward compatibility);
    TRC20 must be enabled server-side. Amounts are always human USDT strings —
    the gateway handles each chain's decimals (BEP20=18, TRC20=6) internally.
servers:
  - url: https://gateway.example.com/api/v1
  - url: http://localhost:4000/api/v1
...
components:
  securitySchemes:
    # Two API key modes coexist. Which one applies is decided by the key itself
    # (api_keys.auth_mode), not by the caller — presenting an HMAC key without a
    # signature, or a bearer token with one, is rejected rather than downgraded.
    #
    # 1. HMAC-signed: ApiKey carries the PUBLIC id (pk_live_…), and Signature +
    #    Timestamp must accompany it. The secret never crosses the wire.
    #
    #    X-Timestamp is unix SECONDS. The signed string is the timestamp, a
    #    literal ".", then the exact raw request body:
    #        X-Signature = hex( HMAC_SHA256( apiSecret, "${X-Timestamp}.${rawBody}" ) )
    #    For a request with no body (any GET) the raw body is "", so the signed
    #    string is "${X-Timestamp}.". Serialize the body once and send those
    #    exact bytes — re-serializing changes them and invalidates the
    #    signature. Requests more than 300 s from server time are rejected as
    #    replays. Verified against backend/src/middleware/auth.ts.
    ApiKey:      { type: apiKey, in: header, name: X-Api-Key }
    Signature:   { type: apiKey, in: header, name: X-Signature }
    Timestamp:   { type: apiKey, in: header, name: X-Timestamp }
    # 2. Bearer: the X-Api-Key value IS the secret (ak_live_…), sent alone.
    #    Simpler to integrate, but the credential is on the wire every call, so
    #    such keys are never granted the payouts:write scope.
    ApiKeyBearer: { type: apiKey, in: header, name: X-Api-Key }
    BearerAuth:  { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    Subscription:
      type: object
      description: >
        A recurring billing SCHEDULE. Each cycle issues an invoice the customer
        chooses to pay — there is no stored mandate to pull funds from.
      properties:
        id:    { type: string, format: uuid }
        title: { type: string, example: Pro plan }
        description:   { type: string, nullable: true }
        customerName:  { type: string, nullable: true }
        customerEmail: { type: string, nullable: true }
        currency: { type: string, example: INR }
        amount:   { type: string, example: "1499" }
        asset:    { type: string, nullable: true }
        network:  { type: string, nullable: true }
        intervalUnit:  { type: string, enum: [day, week, month, year] }
        intervalCount: { type: integer, example: 1 }
        status:
          type: string
          enum: [active, paused, canceled, needs_attention]
          description: >
            `needs_attention` means the plan fell further behind schedule than
            `maxCyclesBehind` allows, so billing stopped rather than emitting a
            burst of back-dated invoices. Resume it to restart on a clean
            schedule; the missed cycles are not billed.
        nextRunAt:
          type: string
          format: date-time
          description: >
            Advanced by exactly one interval from its own previous value, never
            from now(), so a worker outage does not shift every future billing
            date.
        totalCycles:   { type: integer, nullable: true, description: Null = bill indefinitely }
        cyclesBilled:  { type: integer }
        dueDays:  { type: integer, description: Days from issue to due date on each invoice }
        autoSend: { type: boolean, description: Email each invoice; requires customerEmail }
        lastInvoiceId: { type: string, nullable: true }
        lastRunAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateSubscription:
      type: object
      required: [title, amount, currency, intervalUnit]
      properties:
        title:    { type: string }
        description: { type: string }
        amount:   { type: string, example: "1499" }
        currency: { type: string, example: INR, description: Fiat code; each cycle's crypto amount is quoted when it is paid }
        intervalUnit:  { type: string, enum: [day, week, month, year] }
        intervalCount: { type: integer, default: 1, minimum: 1, maximum: 366 }
        customerName:  { type: string }
        customerEmail: { type: string, format: email }
        asset:    { type: string }
        network:  { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
        startAt:  { type: string, format: date-time, description: When the FIRST invoice is minted. Defaults to now. }
        totalCycles: { type: integer }
        dueDays:  { type: integer, default: 7 }
        autoSend: { type: boolean }
paths:
...
  /pay/{token}:
    get:
      tags: [Payment links]
      summary: "PUBLIC: resolve a checkout link (no authentication)"
      description: >
        Reachable by anyone holding the link URL — the token IS the capability.
        Returns the business name, amount and accepted assets, and NOTHING that
        identifies the merchant further: no id, email, wallet, balance, webhook
        or keys. `usable: false` with `unusableReason` means expired, disabled or
        used up; render that as a dead end rather than an error.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { description: Unknown link, or the merchant is not approved }
        "429": { description: Rate limited }
...
  /subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Subscription' } }
    post:
      tags: [Subscriptions]
      summary: Create a recurring billing plan
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubscription' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        "400": { description: Bad request }
  /subscriptions/{id}:
    get:
      tags: [Subscriptions]
      summary: Retrieve a subscription
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        "404": { description: Not found }
  /subscriptions/{id}/invoices:
    get:
      tags: [Subscriptions]
      summary: The invoices this plan has issued
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
...
  /payouts:
    get:
      tags: [Payouts]
      summary: List payouts
      description: >
        Readable with any credential. Note the asymmetry with POST below:
        reading settlement history is a read, but INITIATING one moves money and
        so is restricted to HMAC-signed keys.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: query, name: page,  schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: OK }
    post:
      tags: [Payouts]
      summary: Request a payout to the configured payout wallet
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: string }
                network:
                  type: string
                  enum: [BEP20, TRC20, ERC20, BTC]
                  default: BEP20
                  description: >
                    Chain to settle on. TRC20 requires a payout_wallet_trc20 to be
                    configured and Tron enabled server-side, else returns 400.
      responses:
        "202": { description: Accepted (queued) }
        "400": { description: "No wallet for network / amount exceeds balance", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /admin/payout:
    post:
      tags: [Admin]
      summary: Manually trigger a payout / settlement
      security: [{ BearerAuth: [] }]
      responses: { "202": { description: Accepted } }
...